// packages/gatsby/seo/seo.types.ts
export interface PageSEOProperties {
  /**
   * The title of the page. This value should only be the
   * title of the page. The SEO component will take
   */
  pageTitle: string;
  /**
   * A general description of the page. If this is not provided
   * the SiteSEOProperties.siteDescription will be defaulted
   * instead.
   *
   * This will display for any social cards, search engine cards, etc...
   * This should be defaulted to the basic description of the app.
   *
   * It's probably within your best interests to updated this per page
   */
  pageDescription?: string;
  /**
   * The image that will display for social cards, search engine cards
   * This image is defaulted to whatever the siteLogo is inside of the
   * SiteSEOProperties.siteImage
   */
  pageImage?: string;
  pageKeywords?: string[];
  pageAuthor?: string;
}

export interface SiteSEOProperties {
  // The base path of the site. The best practice is to just default it to `/`
  sitePathPrefix: string;
  // The base title of the site
  siteTitle: string;
  // Alternative site title for SEO
  siteTitleAlt: string;
  // The base description of the site
  siteDescription: string;
  // Path to your image you placed in the 'static' folder
  siteImage: string;
  // The domain of the site. no trailing slash
  siteUrl: string;
  // Base headline for th schema.org JSON-LD
  siteHeadline: string;
  // the language used for the site
  siteLanguage: string;
  // needed for facebook
  siteOgLanguage: string;
  // author of the site
  siteAuthor: string;
  // base set of keywords that will be included on every page
  siteBaseKeywords?: string[];
  // Social
  // @machineq
  twitterHandle: string;
  facebookProfileUrl?: string;
  linkedInProfileUrl?: string;
}